Create adjacency matrix, graph, and find ind. sets
This code takes in a long form data-set of nodes with distances between them. 
It uses a pre-defined rule to decide whether there is an edge between the nodes based on the distance. 
Using this, it creates an adjacency matrix for the graph. 
It then creates a NetworkX graph object from this matrix. 
Finally, it runs various graph algorithms on this object.